move anndata file into anndata subfolder
