+++ scvi/external/solo/_model.py
@@ -10,7 +10,11 @@ import torch
 from anndata import AnnData
 
 from scvi import _CONSTANTS
-from scvi.data._anndata import _setup_anndata, get_from_registry, transfer_anndata_setup
+from scvi.data.anndata._utils import (
+    _setup_anndata,
+    get_from_registry,
+    transfer_anndata_setup,
+)
 from scvi.dataloaders import DataSplitter
 from scvi.model import SCVI
 from scvi.model.base import BaseModelClass
