+++ scvi/model/_totalvi.py
@@ -12,9 +12,9 @@ from anndata import AnnData
 from scvi import _CONSTANTS
 from scvi._compat import Literal
 from scvi._utils import _doc_params
-from scvi.data import get_from_registry
-from scvi.data._anndata import _setup_anndata
 from scvi.data._utils import _check_nonnegative_integers
+from scvi.data.anndata import get_from_registry
+from scvi.data.anndata._utils import _setup_anndata
 from scvi.dataloaders import DataSplitter
 from scvi.model._utils import (
     _get_batch_code_from_category,
