+++ scvi/model/_utils.py
@@ -9,7 +9,7 @@ import scipy.sparse as sp_sparse
 import torch
 
 from scvi import _CONSTANTS
-from scvi.data import get_from_registry
+from scvi.data.anndata import get_from_registry
 
 logger = logging.getLogger(__name__)
