+++ scvi/model/base/_differential.py
@@ -11,7 +11,7 @@ from sklearn.mixture import GaussianMixture
 
 from scvi import _CONSTANTS
 from scvi._compat import Literal
-from scvi.data import get_from_registry
+from scvi.data.anndata import get_from_registry
 
 logger = logging.getLogger(__name__)
